+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Implement using function calls as values](https://github.com/amp-lang/ampc/issues/8)
 - [Implement implicit returns](https://github.com/amp-lang/ampc/issues/6)
 - Report diagnostic instead of panicking for mismatched return types
+- Report diagnostic for function argument number length mismatches
 
 ### [`snapshot-02-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-02-06-2023-a)
 (initial snapshot)